everything works backend so far

+++ Repositories/VaultKeepsRepository.cs
@@ -3,12 +3,12 @@ using System.Data;
 using Dapper;
 using Keepr.Models;
 
-namespace libraryapi.Repositories
+namespace keepr.Repositories
 {
-  public class VaultKeepRepository
+  public class VaultKeepsRepository
   {
     private readonly IDbConnection _db;
-    public VaultKeepRepository(IDbConnection db)
+    public VaultKeepsRepository(IDbConnection db)
     {
       _db = db;
 
@@ -20,28 +20,28 @@ namespace libraryapi.Repositories
     {
       return _db.Query<Keep>($@"
         SELECT * FROM vaultkeeps vk
-        INNER JOIN keep k ON k.id = vk.keepId
+        INNER JOIN keeps k ON k.id = vk.keepId
         WHERE (vaultId = @id);
       ", new { id });
     }
 
     //GetVaultsByKeepId
-    public IEnumerable<Vault> GetVaultsByKeepId(int id)
-    {
-      return _db.Query<Vault>($@"
-        SELECT * FROM vaultkeeps vk
-        INNER JOIN vault v ON v.id = vk.vaultId
-        WHERE (keepId = @id);
-      ", new { id });
-    }
+    // public IEnumerable<Vault> GetVaultsByKeepId(int id)
+    // {
+    //   return _db.Query<Vault>($@"
+    //     SELECT * FROM vaultkeeps vk
+    //     INNER JOIN vault v ON v.id = vk.vaultId
+    //     WHERE (keepId = @id);
+    //   ", new { id });
+    // }
 
 
     //AddVaultKeep
     public VaultKeep AddVaultKeep(VaultKeep vk)
     {
       int id = _db.ExecuteScalar<int>(@"
-      INSERT INTO VaultKeeps(keepId, vaultId)
-      VALUES(@KeepId, @VaultId);
+      INSERT INTO VaultKeeps(keepId, vaultId, userId)
+      VALUES(@KeepId, @VaultId, @UserId);
       SELECT LAST_INSERT_ID();
       ", vk);
       vk.Id = id;
@@ -52,7 +52,7 @@ namespace libraryapi.Repositories
 
     public bool DeleteVaultKeep(VaultKeep vk)
     {
-      int success = _db.Execute(@"DELETE FROM VaultKeeps WHERE keepId = @KeepId AND vaultId = @VaultId", vk);
+      int success = _db.Execute(@"DELETE FROM VaultKeeps WHERE keepId = @KeepId AND vaultId = @VaultId AND userId = @UserId", vk);
       return success != 0;
 
     }
